backend/requirements.txt
gunicorn==20.1.0
Django==3.2.16
django-filter==23.1
djangorestframework==3.12.4
djoser==2.1.0
dotenv==1.0.1
Pillow==9.0.0